feat: add functions.js file & move common functions about signin, signup
