+++ Linkbrary/signin.html
@@ -55,5 +55,6 @@
       </div>
     </div>
     <script type="module" src="./src/signin.js"></script>
+    <script type="module" src="./src/functions.js"></script>
   </body>
 </html>
