+++ Linkbrary/signup.html
@@ -63,5 +63,6 @@
       </div>
     </div>
     <script type="module" src="./src/signup.js"></script>
+    <script type="module" src="./src/functions.js"></script>
   </body>
 </html>
