+++ Linkbrary/src/functions.js
@@ -1,0 +1,51 @@
+import {
+  setInputError,
+  removeInputError,
+  isEmailValid,
+  isPasswordValid,
+  togglePassword,
+  TEST_USER,
+} from "./utils.js";
+
+const emailInput = document.querySelector("#email");
+const emailErrorMessage = document.querySelector("#email-error-message");
+emailInput.addEventListener("focusout", (event) => validateEmailInput(event.target.value));
+function validateEmailInput(email) {
+  if (email === "") {
+    setInputError({ input: emailInput, errorMessage: emailErrorMessage }, "이메일을 입력해주세요.");
+    return;
+  }
+  if (!isEmailValid(email)) {
+    setInputError(
+      { input: emailInput, errorMessage: emailErrorMessage },
+      "올바른 이메일 주소가 아닙니다."
+    );
+    return;
+  }
+}
+
+const passwordInput = document.querySelector("#password");
+const passwordErrorMessage = document.querySelector("#password-error-message");
+passwordInput.addEventListener("focusout", (event) => validatePasswordInput(event.target.value));
+function validatePasswordInput(password) {
+  if (password === "") {
+    setInputError(
+      { input: passwordInput, errorMessage: passwordErrorMessage },
+      "비밀번호를 입력해주세요."
+    );
+    return;
+  }
+  if (!isPasswordValid(password)) {
+    setInputError(
+      { input: passwordInput, errorMessage: passwordErrorMessage },
+      "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요."
+    );
+    return;
+  }
+  removeInputError({ input: passwordInput, errorMessage: passwordErrorMessage });
+}
+
+const passwordToggleButton = document.querySelector("#password-toggle");
+passwordToggleButton.addEventListener("click", () =>
+  togglePassword(passwordInput, passwordToggleButton)
+);
