+++ Linkbrary/src/signup.js
@@ -9,19 +9,8 @@ import {
 
 const emailInput = document.querySelector("#email");
 const emailErrorMessage = document.querySelector("#email-error-message");
-emailInput.addEventListener("focusout", (event) => validateEmailInput(event.target.value));
-function validateEmailInput(email) {
-  if (email === "") {
-    setInputError({ input: emailInput, errorMessage: emailErrorMessage }, "이메일을 입력해주세요.");
-    return;
-  }
-  if (!isEmailValid(email)) {
-    setInputError(
-      { input: emailInput, errorMessage: emailErrorMessage },
-      "올바른 이메일 주소가 아닙니다."
-    );
-    return;
-  }
+emailInput.addEventListener("focusout", (event) => validateEmailDuplicate(event.target.value));
+function validateEmailDuplicate(email) {
   if (email === TEST_USER.email){
     setInputError(
       { input: emailInput, errorMessage: emailErrorMessage },
@@ -30,27 +19,6 @@ function validateEmailInput(email) {
     return;
   }
   removeInputError({ input: emailInput, errorMessage: emailErrorMessage });
-}
-
-const passwordInput = document.querySelector("#password");
-const passwordErrorMessage = document.querySelector("#password-error-message");
-passwordInput.addEventListener("focusout", (event) => validatePasswordInput(event.target.value));
-function validatePasswordInput(password) {
-  if (password === "") {
-    setInputError(
-      { input: passwordInput, errorMessage: passwordErrorMessage },
-      "비밀번호를 입력해주세요."
-    );
-    return;
-  }
-  if (!isPasswordValid(password)) {
-    setInputError(
-      { input: passwordInput, errorMessage: passwordErrorMessage },
-      "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요."
-    );
-    return;
-  }
-  removeInputError({ input: passwordInput, errorMessage: passwordErrorMessage });
 }
 
 const passwordCheckInput = document.querySelector("#password-check");
@@ -69,12 +37,6 @@ function validatePasswordCheckInput(password) {
   }
   removeInputError({ input: passwordCheckInput, errorMessage: passwordCheckErrorMessage });
 }
-
-
-const passwordToggleButton = document.querySelector("#password-toggle");
-passwordToggleButton.addEventListener("click", () =>
-  togglePassword(passwordInput, passwordToggleButton)
-);
 
 const passwordCheckToggleButton = document.querySelector("#password-check-toggle");
 passwordCheckToggleButton.addEventListener("click", () =>
@@ -99,4 +61,6 @@ function submitForm(event) {
     { input: passwordInput, errorMessage: passwordErrorMessage },
     "비밀번호를 확인해주세요."
   );
+  setInputError({ input: passwordCheckInput, errorMessage: passwordCheckErrorMessage },
+    "비밀번호가 일치한지 확인해주세요." );
 }
